# docs/swagger.yaml
basePath: /
definitions:
  handler.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  service.CreateFacilityInput:
    properties:
      client_code_prefix:
        type: string
      code:
        type: string
      district:
        type: string
      hsd:
        type: string
      level:
        type: string
      name:
        type: string
      subcounty:
        type: string
    type: object
...
  utils.APIError:
    properties:
      errors:
        items:
          $ref: '#/definitions/utils.ErrorDetail'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  utils.APIResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      success:
        type: boolean
    type: object
  utils.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  utils.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
info:
  contact: {}
  description: Uganda MoH HMIS MCH 007 Integrated Family Planning Register
  title: HMIS MCH 007 – Integrated FP Register API
  version: "1.0"
paths:
...
  /api/v1/auth/change-password:
    post:
      consumes:
      - application/json
      description: Allows the currently authenticated user to change their password
      parameters:
      - description: Password change payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Change own password
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Log in
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke the refresh token
      parameters:
      - description: Refresh token to revoke
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Log out
      tags:
      - auth
  /api/v1/auth/me:
    get:
      description: Returns the currently authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a refresh token for a new access + refresh token pair
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Refresh tokens
      tags:
      - auth
  /api/v1/facilities:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      - description: Search by name, code, district, or subcounty
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List facilities
      tags:
      - facilities
    post:
      consumes:
      - application/json
      parameters:
      - description: Facility data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateFacilityInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a facility
      tags:
      - facilities
  /api/v1/facilities/{id}:
    delete:
      parameters:
      - description: Facility ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete facility
      tags:
      - facilities
    get:
      parameters:
      - description: Facility ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get facility by ID
      tags:
      - facilities
    put:
      consumes:
      - application/json
      parameters:
      - description: Facility ID
        in: path
        name: id
        required: true
        type: string
      - description: Facility data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateFacilityInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update facility
      tags:
      - facilities
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"